add separators
