+++ docs/concepts/task/remote-cmd-task.md
@@ -363,7 +363,8 @@ __Arguments:__
 
 - `envs` (`Env`): One or more environment variable instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Env
 task = Task(name='task')
@@ -384,7 +385,8 @@ __Arguments:__
 
 - `env_files` (`EnvFile`): One or more environment file instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, EnvFile
 task = Task()
@@ -404,7 +406,8 @@ __Arguments:__
 
 - `inputs` (`AnyInput`): One or more input instances to be added to the input list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Input
 task = Task(name='task')
@@ -424,7 +427,8 @@ __Arguments:__
 
 - `upstreams` (`TAnyTask`): One or more task instances to be added to the upstream list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 task = Task(name='task')
@@ -447,7 +451,8 @@ __Returns:__
 
 `bool`: True if the task is completed, False otherwise.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -471,7 +476,8 @@ __Returns:__
 
 `TAnyTask`: A copy of the current task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 task = Task(name='my-task', cmd='echo hello')
@@ -554,7 +560,8 @@ This method allows for the addition of custom validation or condition checkers. 
 checkers can be used to verify certain conditions before the task execution proceeds.
 Subclasses should implement this method to define task-specific checkers.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -567,7 +574,8 @@ class MyTask(Task):
 
 Injects additional `EnvFile` into the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -580,7 +588,8 @@ class MyTask(Task):
 
 Injects environment variables into the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -597,7 +606,8 @@ This method is used to programmatically add input parameters to the task, allowi
 dynamic customization of the task's input data. Subclasses should override this method
 to define specific inputs that the task should receive.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Input
 class MyTask(Task):
@@ -614,7 +624,8 @@ This method is used for programmatically adding upstream dependencies to the tas
 Upstream tasks are those that must be completed before the current task starts.
 Override this method in subclasses to specify such dependencies.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -634,7 +645,8 @@ __Arguments:__
 
 - `envs` (`Env`): One or more environment variable instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Env
 task = Task(name='task')
@@ -655,7 +667,8 @@ __Arguments:__
 
 - `env_files` (`EnvFile`): One or more environment file instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, EnvFile
 task = Task()
@@ -675,7 +688,8 @@ __Arguments:__
 
 - `inputs` (`AnyInput`): One or more input instances to be added to the input list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Input
 task = Task(name='task')
@@ -696,7 +710,8 @@ __Arguments:__
 
 - `upstreams` (`TAnyTask`): One or more task instances to be added to the upstream list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 task = Task(name='task')
@@ -743,7 +758,8 @@ __Arguments:__
 - `is_last_attempt` (`bool`): Indicates if this is the final retry attempt.
 - `exception` (`Exception`): The exception that caused the task to fail.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -763,7 +779,8 @@ This asynchronous method should be implemented in subclasses to specify
 actions that occur when the task reaches the `ready` state. This can include
 any cleanup, notification, or follow-up actions specific to the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -780,7 +797,8 @@ Implement this method to specify behavior when the task is retried after a failu
 This could include resetting states, logging the retry attempt, or other necessary
 steps before re-execution.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -796,7 +814,8 @@ Defines actions to perform when the task status is set to `skipped`.
 Implement this method to specify behavior when the task is skipped. This could
 include logging information, cleaning up resources, or any other necessary steps.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -812,7 +831,8 @@ Defines actions to perform when the task status is set to 'started'.
 Implement this method to specify behavior when the task starts its execution. This
 could involve initializing resources, logging, or other startup procedures.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -829,7 +849,8 @@ Implement this method to specify behavior when the task transitions to the
 `triggered` state. This could involve setting up prerequisites or sending
 notifications.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -846,7 +867,8 @@ Implement this method to specify behavior when the task transitions to the
 `waiting` state. This state usually indicates the task is waiting for some
 condition or prerequisite to be met.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -882,7 +904,8 @@ __Arguments:__
 
 - `result` (`Any`): The result of the task to be printed.
 
-Example:
+__Examples:__
+
 >> from zrb import Task
 >> # Example of overriding in a subclass
 >> class MyTask(Task):
@@ -937,7 +960,8 @@ __Returns:__
 `Any`: The result of the task execution, the type of which is determined by
 the specific task implementation.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -1055,15 +1079,17 @@ __Returns:__
 
 `Callable[..., Any]`: A callable representation of the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
     async def run(self, *args: Any, **kwargs: Any) -> int:
         self.print_out('Doing some calculation')
         return 42
-````
+```
 
+```
 >>>
 ```python
 task = MyTask()
