+++ docs/concepts/task/resource-maker.md
@@ -395,7 +395,8 @@ __Arguments:__
 
 - `envs` (`Env`): One or more environment variable instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Env
 task = Task(name='task')
@@ -416,7 +417,8 @@ __Arguments:__
 
 - `env_files` (`EnvFile`): One or more environment file instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, EnvFile
 task = Task()
@@ -436,7 +438,8 @@ __Arguments:__
 
 - `inputs` (`AnyInput`): One or more input instances to be added to the input list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Input
 task = Task(name='task')
@@ -456,7 +459,8 @@ __Arguments:__
 
 - `upstreams` (`TAnyTask`): One or more task instances to be added to the upstream list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 task = Task(name='task')
@@ -479,7 +483,8 @@ __Returns:__
 
 `bool`: True if the task is completed, False otherwise.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -503,7 +508,8 @@ __Returns:__
 
 `TAnyTask`: A copy of the current task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 task = Task(name='my-task', cmd='echo hello')
@@ -586,7 +592,8 @@ This method allows for the addition of custom validation or condition checkers. 
 checkers can be used to verify certain conditions before the task execution proceeds.
 Subclasses should implement this method to define task-specific checkers.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -599,7 +606,8 @@ class MyTask(Task):
 
 Injects additional `EnvFile` into the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -612,7 +620,8 @@ class MyTask(Task):
 
 Injects environment variables into the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -629,7 +638,8 @@ This method is used to programmatically add input parameters to the task, allowi
 dynamic customization of the task's input data. Subclasses should override this method
 to define specific inputs that the task should receive.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Input
 class MyTask(Task):
@@ -646,7 +656,8 @@ This method is used for programmatically adding upstream dependencies to the tas
 Upstream tasks are those that must be completed before the current task starts.
 Override this method in subclasses to specify such dependencies.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -666,7 +677,8 @@ __Arguments:__
 
 - `envs` (`Env`): One or more environment variable instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Env
 task = Task(name='task')
@@ -687,7 +699,8 @@ __Arguments:__
 
 - `env_files` (`EnvFile`): One or more environment file instances to be added.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, EnvFile
 task = Task()
@@ -707,7 +720,8 @@ __Arguments:__
 
 - `inputs` (`AnyInput`): One or more input instances to be added to the input list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task, Input
 task = Task(name='task')
@@ -728,7 +742,8 @@ __Arguments:__
 
 - `upstreams` (`TAnyTask`): One or more task instances to be added to the upstream list.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 task = Task(name='task')
@@ -775,7 +790,8 @@ __Arguments:__
 - `is_last_attempt` (`bool`): Indicates if this is the final retry attempt.
 - `exception` (`Exception`): The exception that caused the task to fail.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -795,7 +811,8 @@ This asynchronous method should be implemented in subclasses to specify
 actions that occur when the task reaches the `ready` state. This can include
 any cleanup, notification, or follow-up actions specific to the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -812,7 +829,8 @@ Implement this method to specify behavior when the task is retried after a failu
 This could include resetting states, logging the retry attempt, or other necessary
 steps before re-execution.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -828,7 +846,8 @@ Defines actions to perform when the task status is set to `skipped`.
 Implement this method to specify behavior when the task is skipped. This could
 include logging information, cleaning up resources, or any other necessary steps.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -844,7 +863,8 @@ Defines actions to perform when the task status is set to 'started'.
 Implement this method to specify behavior when the task starts its execution. This
 could involve initializing resources, logging, or other startup procedures.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -861,7 +881,8 @@ Implement this method to specify behavior when the task transitions to the
 `triggered` state. This could involve setting up prerequisites or sending
 notifications.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -878,7 +899,8 @@ Implement this method to specify behavior when the task transitions to the
 `waiting` state. This state usually indicates the task is waiting for some
 condition or prerequisite to be met.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -914,7 +936,8 @@ __Arguments:__
 
 - `result` (`Any`): The result of the task to be printed.
 
-Example:
+__Examples:__
+
 >> from zrb import Task
 >> # Example of overriding in a subclass
 >> class MyTask(Task):
@@ -969,7 +992,8 @@ __Returns:__
 `Any`: The result of the task execution, the type of which is determined by
 the specific task implementation.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
@@ -1087,15 +1111,17 @@ __Returns:__
 
 `Callable[..., Any]`: A callable representation of the task.
 
-Example:
+__Examples:__
+
 ```python
 from zrb import Task
 class MyTask(Task):
     async def run(self, *args: Any, **kwargs: Any) -> int:
         self.print_out('Doing some calculation')
         return 42
-````
+```
 
+```
 >>>
 ```python
 task = MyTask()
