+++ src/zrb/builtin/generator/app_generator/template/base/_automate/generate_snake_zrb_meta_template_name/template/_automate/snake_zrb_app_name/local.py
@@ -3,7 +3,7 @@ from zrb.builtin.group import project_group
 from .container import start_snake_zrb_app_name_container
 
 ###############################################################################
-# Task Definitions
+# ⚙️ start-kebab-zrb-task-name
 ###############################################################################
 
 start_snake_zrb_app_name = CmdTask(
