+++ src/zrb/builtin/generator/app_generator/template/build-custom-image/_automate/generate_snake_zrb_meta_template_name/template/_automate/snake_zrb_app_name/local.py
@@ -3,11 +3,11 @@ from zrb.builtin.group import project_group
 from ._common import (
     CURRENT_DIR, APP_DIR, APP_TEMPLATE_ENV_FILE_NAME, local_input, host_input
 )
-from .checker import snake_zrb_app_name_checker
+from ._checker import snake_zrb_app_name_checker
 import os
 
 ###############################################################################
-# Env File Definitions
+# 🌳 Env File Definitions
 ###############################################################################
 
 app_env_file = EnvFile(
@@ -15,7 +15,7 @@ app_env_file = EnvFile(
 )
 
 ###############################################################################
-# Task Definitions
+# ⚙️ start_kebab-zrb-task-name
 ###############################################################################
 
 start_snake_zrb_app_name = CmdTask(
