+++ src/zrb/builtin/generator/app_generator/template/http-port-build-custom-image/_automate/generate_snake_zrb_meta_template_name/template/_automate/snake_zrb_app_name/local.py
@@ -4,11 +4,11 @@ from ._common import (
     CURRENT_DIR, APP_DIR, APP_TEMPLATE_ENV_FILE_NAME, local_input, host_input,
     https_input
 )
-from .checker import snake_zrb_app_name_checker
+from ._checker import snake_zrb_app_name_checker
 import os
 
 ###############################################################################
-# Env File Definitions
+# 🌳 Env File Definitions
 ###############################################################################
 
 app_env_file = EnvFile(
