+++ src/zrb/builtin/generator/fastapp/template/_automate/snake_zrb_app_name/container.py
@@ -1,97 +1,26 @@
-from typing import Mapping, Any, List
-from zrb import DockerComposeTask, Env, ServiceConfig, EnvFile, runner, Task
-from zrb.helper.util import to_kebab_case
+from zrb import DockerComposeTask, runner
 from zrb.builtin.group import project_group
-from ._common import (
-    APP_TEMPLATE_ENV_FILE_NAME, RESOURCE_DIR, MODULES, app_container_checker,
-    rabbitmq_checker, rabbitmq_management_checker, redpanda_console_checker,
-    kafka_outside_checker, kafka_plaintext_checker, pandaproxy_outside_checker,
-    pandaproxy_plaintext_checker, local_input, run_mode_input,
-    enable_monitoring_input, host_input, https_input, app_enable_otel_env
+from ._checker import (
+    app_container_checker, rabbitmq_checker, rabbitmq_management_checker,
+    redpanda_console_checker, kafka_outside_checker, kafka_plaintext_checker,
+    pandaproxy_outside_checker, pandaproxy_plaintext_checker
 )
-from .image import build_snake_zrb_app_name_image, image_input, image_env
-import os
-
-DOCKER_COMPOSE_APP_ENV_FILE_NAME = os.path.join(
-    RESOURCE_DIR, 'docker-compose-app.env'
+from ._input import (
+    local_input, run_mode_input, enable_monitoring_input, host_input,
+    https_input, image_input
 )
-
-###############################################################################
-# Functions
-###############################################################################
-
-
-def setup_runtime_compose_profile(*args: Any, **kwargs: Any) -> str:
-    task: Task = kwargs.get('_task')
-    env_map = task.get_env_map()
-    compose_profiles: List[str] = [
-        kwargs.get('snake_zrb_app_name_run_mode', 'monolith'),
-    ]
-    broker_type = env_map.get('APP_BROKER_TYPE', 'rabbitmq')
-    if broker_type in ['rabbitmq', 'kafka']:
-        compose_profiles.append(broker_type)
-    if kwargs.get('enable_snake_zrb_app_name_monitoring', False):
-        compose_profiles.append('monitoring')
-    compose_profile_str = ','.join(compose_profiles)
-    return f'export COMPOSE_PROFILES={compose_profile_str}'
-
-
-def setup_all_compose_profile(*args: Any, **kwargs: Any) -> str:
-    compose_profiles = [
-        'monitoring',
-        'monolith',
-        'microservices',
-        'kafka',
-        'rabbitmq'
-    ]
-    compose_profile_str = ','.join(compose_profiles)
-    return f'export COMPOSE_PROFILES={compose_profile_str}'
-
-
-def should_execute(*args: Any, **kwargs: Any) -> bool:
-    return kwargs.get('local_snake_zrb_app_name', True)
-
-
-###############################################################################
-# Env File Definitions
-###############################################################################
-
-compose_env_file = EnvFile(
-    path=os.path.join(RESOURCE_DIR, 'docker-compose.env'),
-    prefix='CONTAINER_ZRB_ENV_PREFIX'
+from ._env import app_enable_otel_env, image_env
+from ._env_file import compose_env_file
+from ._helper import (
+    activate_selected_compose_profile, activate_all_compose_profile,
+    should_start_container
 )
+from ._config import RESOURCE_DIR, SERVICE_CONFIGS
+from .image import build_snake_zrb_app_name_image
+
 
 ###############################################################################
-# Compose Service Config Definitions
-###############################################################################
-
-service_configs: Mapping[str, ServiceConfig] = {}
-service_names = [to_kebab_case(module) + '-service' for module in MODULES]
-for suffix in ['', 'gateway'] + service_names:
-    service_suffix = '-' + suffix if suffix != '' else ''
-    service_name = f'kebab-zrb-app-name{service_suffix}'
-    service_configs[service_name] = ServiceConfig(
-        env_files=[
-            EnvFile(
-                path=APP_TEMPLATE_ENV_FILE_NAME,
-                prefix='CONTAINER_ZRB_ENV_PREFIX'
-            ),
-            EnvFile(
-                path=DOCKER_COMPOSE_APP_ENV_FILE_NAME,
-                prefix='CONTAINER_ZRB_ENV_PREFIX'
-            )
-        ],
-        envs=[
-            Env(
-                name='APP_OTEL_EXPORTER_OTLP_ENDPOINT',
-                os_name='',
-                default='http://otel-collector:4317'
-            )
-        ]
-    )
-
-###############################################################################
-# Task Definitions
+# ⚙️ remove-kebab-zrb-task-name-container
 ###############################################################################
 
 remove_snake_zrb_app_name_container = DockerComposeTask(
@@ -100,14 +29,18 @@ remove_snake_zrb_app_name_container = DockerComposeTask(
     description='Remove human readable zrb app name container',
     group=project_group,
     cwd=RESOURCE_DIR,
-    setup_cmd=setup_all_compose_profile,
+    setup_cmd=activate_all_compose_profile,
     compose_cmd='down',
     compose_env_prefix='CONTAINER_ZRB_ENV_PREFIX',
-    compose_service_configs=service_configs,
+    compose_service_configs=SERVICE_CONFIGS,
     envs=[image_env],
     env_files=[compose_env_file]
 )
 runner.register(remove_snake_zrb_app_name_container)
+
+###############################################################################
+# ⚙️ init-kebab-zrb-task-name-container
+###############################################################################
 
 init_snake_zrb_app_name_container = DockerComposeTask(
     icon='🔥',
@@ -120,23 +53,27 @@ init_snake_zrb_app_name_container = DockerComposeTask(
         host_input,
         image_input,
     ],
-    should_execute=should_execute,
+    should_execute=should_start_container,
     upstreams=[
         build_snake_zrb_app_name_image,
         remove_snake_zrb_app_name_container
     ],
     cwd=RESOURCE_DIR,
-    setup_cmd=setup_runtime_compose_profile,
+    setup_cmd=activate_selected_compose_profile,
     compose_cmd='up',
     compose_flags=['-d'],
     compose_env_prefix='CONTAINER_ZRB_ENV_PREFIX',
-    compose_service_configs=service_configs,
+    compose_service_configs=SERVICE_CONFIGS,
     envs=[
         image_env,
         app_enable_otel_env,
     ],
     env_files=[compose_env_file],
 )
+
+###############################################################################
+# ⚙️ start-kebab-zrb-task-name-container
+###############################################################################
 
 start_snake_zrb_app_name_container = DockerComposeTask(
     icon='🐳',
@@ -150,14 +87,14 @@ start_snake_zrb_app_name_container = DockerComposeTask(
         https_input,
         image_input,
     ],
-    should_execute=should_execute,
+    should_execute=should_start_container,
     upstreams=[init_snake_zrb_app_name_container],
     cwd=RESOURCE_DIR,
-    setup_cmd=setup_runtime_compose_profile,
+    setup_cmd=activate_selected_compose_profile,
     compose_cmd='logs',
     compose_flags=['-f'],
     compose_env_prefix='CONTAINER_ZRB_ENV_PREFIX',
-    compose_service_configs=service_configs,
+    compose_service_configs=SERVICE_CONFIGS,
     envs=[image_env],
     env_files=[compose_env_file],
     checkers=[
@@ -173,16 +110,20 @@ start_snake_zrb_app_name_container = DockerComposeTask(
 )
 runner.register(start_snake_zrb_app_name_container)
 
+###############################################################################
+# ⚙️ stop-kebab-zrb-task-name-container
+###############################################################################
+
 stop_snake_zrb_app_name_container = DockerComposeTask(
     icon='⛔',
     name='stop-kebab-zrb-app-name-container',
     description='Stop human readable zrb app name container',
     group=project_group,
     cwd=RESOURCE_DIR,
-    setup_cmd=setup_all_compose_profile,
+    setup_cmd=activate_all_compose_profile,
     compose_cmd='stop',
     compose_env_prefix='CONTAINER_ZRB_ENV_PREFIX',
-    compose_service_configs=service_configs,
+    compose_service_configs=SERVICE_CONFIGS,
     envs=[image_env],
     env_files=[compose_env_file]
 )
