+++ src/zrb/builtin/generator/fastapp/template/_automate/snake_zrb_app_name/deployment.py
@@ -1,89 +1,20 @@
-from zrb import CmdTask, Env, EnvFile, StrInput, ChoiceInput, runner
+from zrb import CmdTask, runner
 from zrb.builtin.group import project_group
 from .image import push_snake_zrb_app_name_image
-from ._common import (
-    CURRENT_DIR, DEPLOYMENT_DIR, APP_TEMPLATE_ENV_FILE_NAME,
-    DEPLOYMENT_TEMPLATE_ENV_FILE_NAME, MODULES,
-    enable_monitoring_input
+from ._config import CURRENT_DIR, DEPLOYMENT_DIR
+from ._input import (
+    enable_monitoring_input, image_input, deploy_mode_input, pulumi_stack_input
 )
-from .image import image_input, image_env
+from ._env import (
+    image_env, pulumi_backend_url_env, pulumi_config_passphrase_env,
+    deployment_modules_env, deployment_mode_env,
+    deployment_enable_monitoring_env
+)
+from ._env_file import deployment_app_env_file, deployment_config_env_file
 import os
-import jsons
-
 
 ###############################################################################
-# Input Definitions
-###############################################################################
-
-deploy_mode_input = ChoiceInput(
-    name='kebab-zrb-app-name-deploy-mode',
-    description='"kebab-zrb-app-name" deploy mode (monolith/microservices)',
-    prompt='Deploy "kebab-zrb-app-name" as a monolith or microservices?',
-    choices=['monolith', 'microservices'],
-    default='monolith'
-)
-
-pulumi_stack_input = StrInput(
-    name='kebab-zrb-app-name-pulumi-stack',
-    description='Pulumi stack name for "kebab-zrb-app-name"',
-    prompt='Pulumi stack name for "kebab-zrb-app-name"',
-    default=os.getenv('ZRB_ENV', 'dev')
-)
-
-###############################################################################
-# Env Definitions
-###############################################################################
-
-pulumi_backend_url_env = Env(
-    name='PULUMI_BACKEND_URL',
-    os_name='PULUMI_ZRB_ENV_PREFIX_BACKEND_URL',
-    default=f'file://{DEPLOYMENT_DIR}/state'
-)
-
-pulumi_config_passphrase_env = Env(
-    name='PULUMI_CONFIG_PASSPHRASE',
-    os_name='PULUMI_ZRB_ENV_PREFIX_CONFIG_PASSPHRASE',
-    default='secret'
-)
-
-###############################################################################
-# Env File Definitions
-###############################################################################
-
-deployment_app_env_file = EnvFile(
-    path=APP_TEMPLATE_ENV_FILE_NAME,
-    prefix='DEPLOYMENT_APP_ZRB_ENV_PREFIX'
-)
-
-deployment_config_env_file = EnvFile(
-    path=DEPLOYMENT_TEMPLATE_ENV_FILE_NAME,
-    prefix='DEPLOYMENT_CONFIG_ZRB_ENV_PREFIX'
-)
-
-###############################################################################
-# Env Definitions
-###############################################################################
-
-deployment_modules_env = Env(
-    name='MODULES',
-    os_name='DEPLOYMENT_CONFIG_ZRB_ENV_PREFIX_MODULES',
-    default=jsons.dumps(MODULES)
-)
-
-deployment_mode_env = Env(
-    name='MODE',
-    os_name='DEPLOYMENT_CONFIG_ZRB_ENV_PREFIX_MODE',
-    default='{{input.snake_zrb_app_name_deploy_mode}}'
-)
-
-deployment_enable_monitoring_env = Env(
-    name='ENABLE_MONITORING',
-    os_name='DEPLOYMENT_CONFIG_ZRB_ENV_PREFIX_ENABLE_MONITORING',
-    default='{{ 1 if input.enable_snake_zrb_app_name_monitoring else 0 }}'
-)
-
-###############################################################################
-# Task Definitions
+# ⚙️ deploy-kebab-zrb-task-name
 ###############################################################################
 
 deploy_snake_zrb_app_name = CmdTask(
@@ -117,6 +48,10 @@ deploy_snake_zrb_app_name = CmdTask(
     ]
 )
 runner.register(deploy_snake_zrb_app_name)
+
+###############################################################################
+# ⚙️ destroy-kebab-zrb-task-name
+###############################################################################
 
 destroy_snake_zrb_app_name = CmdTask(
     icon='💨',
