+++ src/zrb/builtin/generator/fastapp/template/_automate/snake_zrb_app_name/load_test.py
@@ -1,65 +1,12 @@
 from zrb import CmdTask, EnvFile, BoolInput, StrInput, IntInput, runner
 from zrb.builtin.group import project_group
-from ._common import CURRENT_DIR, RESOURCE_DIR
+from ._config import (
+    CURRENT_DIR, LOAD_TEST_DIR, LOAD_TEST_TEMPLATE_ENV_FILE_NAME
+)
 import os
 
 ###############################################################################
-# Constants
-###############################################################################
-
-LOAD_TEST_DIR = os.path.join(RESOURCE_DIR, 'loadtest')
-LOAD_TEST_TEMPLATE_ENV_FILE_NAME = os.path.join(RESOURCE_DIR, 'template.env')
-
-###############################################################################
-# Input Definitions
-###############################################################################
-
-headless_input = BoolInput(
-    name='kebab-zrb-app-name-load-test-headless',
-    default=True,
-    description='Load test UI headless',
-    prompt='Load test UI headless (if True, there will be no UI)',
-)
-
-web_port_input = IntInput(
-    name='kebab-zrb-app-name-load-test-port',
-    default=8089,
-    description='Load test UI web port',
-    prompt='Load test UI web port (Only works if headless is False)',
-)
-
-users_input = IntInput(
-    name='kebab-zrb-app-name-load-test-users',
-    default=200,
-    description='Load test users',
-    prompt='Load test users',
-)
-
-spawn_rate_input = IntInput(
-    name='kebab-zrb-app-name-load-test-spawn-rate',
-    default=10,
-    description='Load test spawn rate',
-    prompt='Load test spawn rate',
-)
-
-url_input = StrInput(
-    name='kebab-zrb-app-name-load-test-url',
-    default='http://localhost:zrbAppHttpPort',
-    description='Load test url',
-    prompt='Load test url',
-)
-
-###############################################################################
-# Env file Definitions
-###############################################################################
-
-load_test_env_file = EnvFile(
-    path=LOAD_TEST_TEMPLATE_ENV_FILE_NAME,
-    prefix='LOAD_TEST_ZRB_ENV_PREFIX'
-)
-
-###############################################################################
-# Task Definitions
+# ⚙️ prepare-kebab-zrb-task-name-load-test
 ###############################################################################
 
 prepare_snake_zrb_app_name_load_test = CmdTask(
@@ -75,6 +22,10 @@ prepare_snake_zrb_app_name_load_test = CmdTask(
 )
 runner.register(prepare_snake_zrb_app_name_load_test)
 
+###############################################################################
+# ⚙️ load-test-kebab-zrb-task-name
+###############################################################################
+
 load_test_snake_zrb_app_name_load_test = CmdTask(
     icon='🧪',
     name='load-test-kebab-zrb-app-name',
@@ -82,14 +33,44 @@ load_test_snake_zrb_app_name_load_test = CmdTask(
     group=project_group,
     upstreams=[prepare_snake_zrb_app_name_load_test],
     inputs=[
-        headless_input,
-        web_port_input,
-        users_input,
-        spawn_rate_input,
-        url_input,
+        BoolInput(
+            name='kebab-zrb-app-name-load-test-headless',
+            default=True,
+            description='Load test UI headless',
+            prompt='Load test UI headless (if True, there will be no UI)',
+        ),
+        IntInput(
+            name='kebab-zrb-app-name-load-test-port',
+            default=8089,
+            description='Load test UI web port',
+            prompt='Load test UI web port (Only works if headless is False)',
+        ),
+        IntInput(
+            name='kebab-zrb-app-name-load-test-users',
+            default=200,
+            description='Load test users',
+            prompt='Load test users',
+        ),
+        IntInput(
+            name='kebab-zrb-app-name-load-test-spawn-rate',
+            default=10,
+            description='Load test spawn rate',
+            prompt='Load test spawn rate',
+        ),
+        StrInput(
+            name='kebab-zrb-app-name-load-test-url',
+            default='http://localhost:zrbAppHttpPort',
+            description='Load test url',
+            prompt='Load test url',
+        )
     ],
     cwd=LOAD_TEST_DIR,
-    env_files=[load_test_env_file],
+    env_files=[
+        EnvFile(
+            path=LOAD_TEST_TEMPLATE_ENV_FILE_NAME,
+            prefix='LOAD_TEST_ZRB_ENV_PREFIX'
+        )
+    ],
     cmd_path=[
         os.path.join(CURRENT_DIR, 'cmd', 'activate-venv.sh'),
         os.path.join(CURRENT_DIR, 'cmd', 'app-load-test.sh'),
