+++ src/zrb/builtin/generator/pip_package/template/_automate/snake_zrb_package_name/local.py
@@ -9,7 +9,7 @@ RESOURCE_DIR = os.path.join(PROJECT_DIR, 'src', 'kebab-zrb-package-name')
 PACKAGE_DIR = os.path.join(RESOURCE_DIR, 'src')
 
 ###############################################################################
-# Task Definitions
+# ⚙️ prepare-kebab-zrb-task-name
 ###############################################################################
 
 prepare_snake_zrb_package_name = CmdTask(
@@ -24,6 +24,10 @@ prepare_snake_zrb_package_name = CmdTask(
 )
 runner.register(prepare_snake_zrb_package_name)
 
+###############################################################################
+# ⚙️ build-kebab-zrb-task-name
+###############################################################################
+
 build_snake_zrb_package_name = CmdTask(
     name='build-kebab-zrb-package-name',
     description='Build human readable zrb package name',
@@ -36,6 +40,10 @@ build_snake_zrb_package_name = CmdTask(
     ]
 )
 runner.register(build_snake_zrb_package_name)
+
+###############################################################################
+# ⚙️ publish-kebab-zrb-task-name
+###############################################################################
 
 publish_snake_zrb_package_name = CmdTask(
     name='publish-kebab-zrb-package-name',
@@ -57,6 +65,10 @@ publish_snake_zrb_package_name = CmdTask(
     ]
 )
 runner.register(publish_snake_zrb_package_name)
+
+###############################################################################
+# ⚙️ install-kebab-zrb-task-name-symlink
+###############################################################################
 
 install_snake_zrb_package_name_symlink = CmdTask(
     name='install-kebab-zrb-package-name-symlink',
