+++ src/zrb/builtin/generator/python_task/template/_automate/snake_zrb_task_name.py
@@ -3,7 +3,7 @@ from zrb import Task, python_task, runner
 from zrb.builtin.group import project_group
 
 ###############################################################################
-# Task Definitions
+# ⚙️ kebab-zrb-task-name
 ###############################################################################
