+++ src/zrb/builtin/generator/simple_python_app/template/_automate/snake_zrb_app_name/deployment.py
@@ -8,7 +8,7 @@ from ._common import (
 import os
 
 ###############################################################################
-# Input Definitions
+# 🔤 Input Definitions
 ###############################################################################
 
 replica_input = IntInput(
@@ -26,7 +26,7 @@ pulumi_stack_input = StrInput(
 )
 
 ###############################################################################
-# Env File Definitions
+# 🌳 Env File Definitions
 ###############################################################################
 
 deployment_app_env_file = EnvFile(
@@ -40,7 +40,7 @@ deployment_config_env_file = EnvFile(
 )
 
 ###############################################################################
-# Env Definitions
+# 🌱 Env Definitions
 ###############################################################################
 
 pulumi_backend_url_env = Env(
@@ -62,7 +62,7 @@ deployment_replica_env = Env(
 )
 
 ###############################################################################
-# Task Definitions
+# ⚙️ deploy-kebab-zrb-task-name
 ###############################################################################
 
 deploy_snake_zrb_app_name = CmdTask(
@@ -93,6 +93,10 @@ deploy_snake_zrb_app_name = CmdTask(
     ]
 )
 runner.register(deploy_snake_zrb_app_name)
+
+###############################################################################
+# ⚙️ destroy-kebab-zrb-task-name
+###############################################################################
 
 destroy_snake_zrb_app_name = CmdTask(
     icon='💨',
