+++ zrb_init.py
@@ -34,7 +34,7 @@ playground_group = Group(
 )
 
 ###############################################################################
-# Input Definitions
+# 🔤 Input Definitions
 ###############################################################################
 
 zrb_version_input = StrInput(
@@ -65,7 +65,7 @@ create_playground_input = BoolInput(
 )
 
 ###############################################################################
-# Env Definitions
+# 🌱 Env Definitions
 ###############################################################################
 
 zrb_version_env = Env(
